Pull out inner classes.
